chore: combine config runner and default runner
