+++ internal/pkg/cli/task_run.go
@@ -186,22 +186,12 @@ func (o *runTaskOpts) configureRunner(sess *awssession.Session) taskRunner {
 		}
 	}
 
-	if o.subnets != nil {
-		return &task.NetworkConfigRunner{
-			Count:     o.count,
-			GroupName: o.groupName,
-
-			Subnets:        o.subnets,
-			SecurityGroups: o.securityGroups,
-
-			ClusterGetter: ecsService,
-			Starter:       ecsService,
-		}
-	}
-
-	return &task.DefaultVPCRunner{
+	return &task.NetworkConfigRunner{
 		Count:     o.count,
 		GroupName: o.groupName,
+
+		Subnets:        o.subnets,
+		SecurityGroups: o.securityGroups,
 
 		VPCGetter:     vpcGetter,
 		ClusterGetter: ecsService,
