+++ internal/pkg/task/config_runner_test.go
@@ -5,7 +5,9 @@ package task
 
 import (
 	"errors"
+	"fmt"
 	"github.com/aws/aws-sdk-go/aws"
+	"github.com/aws/copilot-cli/internal/pkg/aws/ec2"
 	"github.com/aws/copilot-cli/internal/pkg/aws/ecs"
 	"github.com/aws/copilot-cli/internal/pkg/task/mocks"
 	"github.com/golang/mock/gomock"
@@ -23,13 +25,20 @@ func TestNetworkConfigRunner_Run(t *testing.T) {
 
 		mockClusterGetter func(m *mocks.MockDefaultClusterGetter)
 		mockStarter       func(m *mocks.MockTaskRunner)
+		mockVPCGetter     func(m *mocks.MockVPCGetter)
 
 		wantedError error
 		wantedTasks []*Task
 	}{
-		"failed to get clusters": {
+		"failed to get default cluster": {
+			subnets: []string{"subnet-1", "subnet-2"},
+
 			mockClusterGetter: func(m *mocks.MockDefaultClusterGetter) {
 				m.EXPECT().DefaultCluster().Return("", errors.New("error getting default cluster"))
+			},
+			mockVPCGetter: func(m *mocks.MockVPCGetter) {
+				m.EXPECT().SubnetIDs().AnyTimes()
+				m.EXPECT().SecurityGroups().AnyTimes()
 			},
 			mockStarter: func(m *mocks.MockTaskRunner) {
 				m.EXPECT().RunTask(gomock.Any()).Times(0)
@@ -38,7 +47,7 @@ func TestNetworkConfigRunner_Run(t *testing.T) {
 				parentErr: errors.New("error getting default cluster"),
 			},
 		},
-		"failed to kick off task": {
+		"failed to kick off tasks with input subnets and security groups": {
 			count:     1,
 			groupName: "my-task",
 
@@ -47,6 +56,9 @@ func TestNetworkConfigRunner_Run(t *testing.T) {
 
 			mockClusterGetter: func(m *mocks.MockDefaultClusterGetter) {
 				m.EXPECT().DefaultCluster().Return("cluster-1", nil)
+			},
+			mockVPCGetter: func(m *mocks.MockVPCGetter) {
+				m.EXPECT().SubnetIDs([]ec2.Filter{ec2.FilterForDefaultVPCSubnets}).Times(0)
 			},
 			mockStarter: func(m *mocks.MockTaskRunner) {
 				m.EXPECT().RunTask(gomock.Any()).Return(nil, errors.New("error running task"))
@@ -67,11 +79,60 @@ func TestNetworkConfigRunner_Run(t *testing.T) {
 			mockClusterGetter: func(m *mocks.MockDefaultClusterGetter) {
 				m.EXPECT().DefaultCluster().Return("cluster-1", nil)
 			},
+			mockVPCGetter: func(m *mocks.MockVPCGetter) {
+				m.EXPECT().SubnetIDs([]ec2.Filter{ec2.FilterForDefaultVPCSubnets}).Times(0)
+			},
 			mockStarter: func(m *mocks.MockTaskRunner) {
 				m.EXPECT().RunTask(ecs.RunTaskInput{
 					Cluster:        "cluster-1",
 					Count:          1,
 					Subnets:        []string{"subnet-1", "subnet-2"},
+					SecurityGroups: []string{"sg-1", "sg-2"},
+					TaskFamilyName: taskFamilyName("my-task"),
+					StartedBy:      startedBy,
+				}).Return([]*ecs.Task{
+					{
+						TaskArn: aws.String("task-1"),
+					},
+				}, nil)
+			},
+
+			wantedTasks: []*Task{
+				{
+					TaskARN: "task-1",
+				},
+			},
+		},
+		"failed to get default subnets": {
+			mockClusterGetter: func(m *mocks.MockDefaultClusterGetter) {
+				m.EXPECT().DefaultCluster().AnyTimes()
+			},
+			mockVPCGetter: func(m *mocks.MockVPCGetter) {
+				m.EXPECT().SubnetIDs([]ec2.Filter{ec2.FilterForDefaultVPCSubnets}).Return(nil, errors.New("error getting subnets"))
+			},
+			mockStarter: func(m *mocks.MockTaskRunner) {
+				m.EXPECT().RunTask(gomock.Any()).Times(0)
+			},
+			wantedError: fmt.Errorf(fmtErrDefaultSubnets, errors.New("error getting subnets")),
+		},
+		"successfully kick off task with default subnets": {
+			count:     1,
+			groupName: "my-task",
+
+			securityGroups: []string{"sg-1", "sg-2"},
+
+			mockClusterGetter: func(m *mocks.MockDefaultClusterGetter) {
+				m.EXPECT().DefaultCluster().Return("cluster-1", nil)
+			},
+			mockVPCGetter: func(m *mocks.MockVPCGetter) {
+				m.EXPECT().SubnetIDs([]ec2.Filter{ec2.FilterForDefaultVPCSubnets}).
+					Return([]string{"default-subnet-1", "default-subnet-2"}, nil)
+			},
+			mockStarter: func(m *mocks.MockTaskRunner) {
+				m.EXPECT().RunTask(ecs.RunTaskInput{
+					Cluster:        "cluster-1",
+					Count:          1,
+					Subnets:        []string{"default-subnet-1", "default-subnet-2"},
 					SecurityGroups: []string{"sg-1", "sg-2"},
 					TaskFamilyName: taskFamilyName("my-task"),
 					StartedBy:      startedBy,
@@ -95,9 +156,11 @@ func TestNetworkConfigRunner_Run(t *testing.T) {
 			ctrl := gomock.NewController(t)
 			defer ctrl.Finish()
 
+			mockVpcGetter := mocks.NewMockVPCGetter(ctrl)
 			mockClusterGetter := mocks.NewMockDefaultClusterGetter(ctrl)
 			mockStarter := mocks.NewMockTaskRunner(ctrl)
 
+			tc.mockVPCGetter(mockVpcGetter)
 			tc.mockClusterGetter(mockClusterGetter)
 			tc.mockStarter(mockStarter)
 
@@ -108,6 +171,7 @@ func TestNetworkConfigRunner_Run(t *testing.T) {
 				Subnets:        tc.subnets,
 				SecurityGroups: tc.securityGroups,
 
+				VPCGetter:     mockVpcGetter,
 				ClusterGetter: mockClusterGetter,
 				Starter:       mockStarter,
 			}
